Issue #68  divide deleting into blocks by id and path

+++ sql/oracle/I2B2_DELETE_ALL_DATA.sql
@@ -35,6 +35,7 @@ AS
   stepCt number(18,0);
   more_trial exception;
   path_not_found exception;
+  arg_not_valid exception;
 	res	number;
 
 BEGIN
@@ -56,6 +57,10 @@ BEGIN
 
   stepCt := stepCt + 1;
   cz_write_audit(jobId, databaseName, procedureName,'Starting I2B2_DELETE_ALL_DATA', 0, stepCt, 'Done');
+
+  if (path_string is null and trial_id is null) then
+    RAISE arg_not_valid;
+  end if;
 
   if (path_string is not null) then
 		pathString := REGEXP_REPLACE('\' || path_string || '\','(\\){2,}', '\');
@@ -84,56 +89,40 @@ BEGIN
   end if;
 
   if (path_string is null) then
+    BEGIN
     SELECT DISTINCT
       first_value(concept_path) over (partition by sourcesystem_cd order by concept_path)
     INTO pathString
     FROM i2b2demodata.concept_dimension
     WHERE sourcesystem_cd = TrialID;
-
-    If pathString is NULL THEN
-      RAISE path_not_found;
-    END IF;
+    EXCEPTION
+    WHEN NO_DATA_FOUND THEN
+      pathString := NULL;
+      stepCt := stepCt + 1;
+	    cz_write_audit(jobId, databaseName, procedureName, 'Path string not found', 0, stepCt, 'Done');
+    END;
   end if;
 
-	stepCt := stepCt + 1;
-	cz_write_audit(jobId, databaseName, procedureName, 'Path: ' || pathString, 0, stepCt, 'Done');
-
-  select count(parent_concept_path) into topNodeCount
-    from I2B2DEMODATA.concept_counts
-    where
-    concept_path = pathString;
-
-  if (topNodeCount > 0) then
-    select distinct parent_concept_path into topNode
-      from I2B2DEMODATA.concept_counts
-      where
-      concept_path = pathString;
-  else
-    topNode := SUBSTR(pathString,0,instr(pathString,'\',2));
-  end if;
-
-  if pathString != '' or pathString != '%'
-  then
-		stepCt := stepCt + 1;
+  if (pathString is not null and (pathString != '' or pathString != '%')) then
+	  stepCt := stepCt + 1;
 		cz_write_audit(jobId, databaseName, procedureName, 'Deleting data for path: ' || pathString, 0, stepCt, 'Done');
 
-		COMMIT;
-
 		--	delete all i2b2 nodes
-
 		i2b2_delete_all_nodes(pathString,jobId);
 
 		--	delete any table_access data
 		delete from table_access
 		where c_fullname like pathString || '%';
 
-	--	delete any i2b2_tag data
+	  --	delete any i2b2_tag data
+	  delete from i2b2_tags
+	  where path like pathString || '%';
+	  stepCt := stepCt + 1;
+	  cz_write_audit(jobId,databaseName,procedureName,'Delete data for trial from I2B2METADATA i2b2_tags',SQL%ROWCOUNT,stepCt,'Done');
+	  commit;
 
-	delete from i2b2_tags
-	where path like pathString || '%';
-	stepCt := stepCt + 1;
-	cz_write_audit(jobId,databaseName,procedureName,'Delete data for trial from I2B2METADATA i2b2_tags',SQL%ROWCOUNT,stepCt,'Done');
-	commit;
+	  i2b2_remove_empty_parent_nodes(pathString, jobID);
+	end if;
 
 	--	delete clinical data
 	if (trialId is not NUll)
@@ -312,17 +301,14 @@ BEGIN
       cz_write_audit(jobId,databaseName,procedureName,'Delete RNASeq data for trial from DEAPP de_subject_rbm_data',SQL%ROWCOUNT,stepCt,'Done');
       commit;
     end if;
-	end if;
 
-		delete from deapp.de_subject_snp_dataset
+    delete from deapp.de_subject_snp_dataset
 		where trial_name = trialId;
 		stepCt := stepCt + 1;
 		cz_write_audit(jobId,databaseName,procedureName,'Delete data for trial from DE_SUBJECT_SNP_DATASET',SQL%ROWCOUNT,stepCt,'Done');
 		commit;
 
-		i2b2_remove_empty_parent_nodes(pathString, jobID);
-
-  end if;
+	end if;
 
     ---Cleanup OVERALL JOB if this proc is being run standalone
   IF newJobFlag = 1
